# core/types: faster RLP encoding of Header, StateAcccount, ReceiptForStorage