--- a/core/types/block_test.go
+++ b/core/types/block_test.go
@@ -285,7 +285,7 @@ func makeBenchBlock() *Block {
 func TestRlpDecodeParentHash(t *testing.T) {
 	// A minimum one
 	want := common.HexToHash("0x112233445566778899001122334455667788990011223344556677889900aabb")
-	if rlpData, err := rlp.EncodeToBytes(Header{ParentHash: want}); err != nil {
+	if rlpData, err := rlp.EncodeToBytes(&Header{ParentHash: want}); err != nil {
 		t.Fatal(err)
 	} else {
 		if have := HeaderParentHashFromRLP(rlpData); have != want {
@@ -299,7 +299,7 @@ func TestRlpDecodeParentHash(t *testing.T) {
 	// | BaseFee     | dynamic| *big.Int       | 64 bits               |
 	mainnetTd := new(big.Int)
 	mainnetTd.SetString("5ad3c2c71bbff854908", 16)
-	if rlpData, err := rlp.EncodeToBytes(Header{
+	if rlpData, err := rlp.EncodeToBytes(&Header{
 		ParentHash: want,
 		Difficulty: mainnetTd,
 		Number:     new(big.Int).SetUint64(math.MaxUint64),
@@ -316,7 +316,7 @@ func TestRlpDecodeParentHash(t *testing.T) {
 	{
 		// The rlp-encoding of the heder belowCauses _total_ length of 65540,
 		// which is the first to blow the fast-path.
-		h := Header{
+		h := &Header{
 			ParentHash: want,
 			Extra:      make([]byte, 65041),
 		}