@fjl 2022-02-18
These changes were already submitted/reviewed in #24251, but I removed them before merging to get the core/types changes into a separate commit.